refactor : 퀴즈 생성 컴포넌트 1차 분리
